v1.41更新
- 新特性:可选择是否在退出时弹窗.(可选,附加设置,默认启用)
- 新特性:根据兼容性测试结果自动判断是否启用[自动检测Alpha通道].
- 新特性:保存视频缓存.(可选,附加设置,默认启用)
- 新特性:在用户停止任务或者发生错误时,视频的缓存不会被自动删除,并且还可以被软件读取以继续之前的进度.

+++ SRC/Waifu2x-Extension-QT/Anime4K.cpp
@@ -27,17 +27,17 @@
 int MainWindow::Anime4k_Video(int rowNum)
 {
     //============================= 读取设置 ================================
-    int ScaleRatio = ui->spinBox_ScaleRatio_video->value();//放大倍数
-    bool DelOriginal = ui->checkBox_DelOriginal->checkState();//是否删除源文件
-    int Sub_video_ThreadNumRunning = 0;//子线程总数
+    int ScaleRatio = ui->spinBox_ScaleRatio_video->value();
+    bool DelOriginal = ui->checkBox_DelOriginal->checkState();
+    bool isCacheExists = false;
+    bool isVideoConfigChanged = true;
+    int Sub_video_ThreadNumRunning = 0;
     QString OutPutPath_Final ="";
     //========================= 拆解map得到参数 =============================
-    //=== 将状态改为正在处理 =====
     QString status = "Processing";
     emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
-    //========================
     QString SourceFile_fullPath = Table_model_video->item(rowNum,2)->text();
-    if(!file_isFileExist(SourceFile_fullPath))//判断源文件是否存在
+    if(!file_isFileExist(SourceFile_fullPath))
     {
         emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [File does not exist.]"));
         status = "Failed";
@@ -68,69 +68,160 @@ int MainWindow::Anime4k_Video(int rowNum)
     {
         file_path = file_path.left(file_path.length() - 1);
     }
-    QString video_mp4_fullpath="";
+    //===================================================================
+    QString video_mp4_fullpath="";//mp4路径
     if(file_ext!="mp4")
     {
         video_mp4_fullpath = file_path+"/"+file_name+"_"+file_ext+".mp4";
-        if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
     }
     else
     {
         video_mp4_fullpath = file_path+"/"+file_name+".mp4";
     }
-    QString AudioPath = file_path+"/audio_"+file_name+"_"+file_ext+"_waifu2x.wav";
-    //============================== 拆分 ==========================================
+    QString AudioPath = file_path+"/audio_"+file_name+"_"+file_ext+"_waifu2x.wav";//音频
     QString SplitFramesFolderPath = file_path+"/"+file_name+"_"+file_ext+"_splitFrames_waifu2x";//拆分后存储frame的文件夹
-    if(file_isDirExist(SplitFramesFolderPath))
+    QString ScaledFramesFolderPath = SplitFramesFolderPath+"/scaled";//存储放大后的帧
+    //==========================
+    //   检测之前的视频配置文件
+    //==========================
+    QString VideoConfiguration_fullPath = file_path+"/VideoConfiguration_"+file_name+"_"+file_ext+"_Waifu2xEX.ini";
+    if(file_isFileExist(VideoConfiguration_fullPath))
     {
-        file_DelDir(SplitFramesFolderPath);
-        file_mkDir(SplitFramesFolderPath);
+        QSettings *configIniRead = new QSettings(VideoConfiguration_fullPath, QSettings::IniFormat);
+        configIniRead->setIniCodec(QTextCodec::codecForName("UTF-8"));
+        //=================== 加载之前存储的视频信息 =========================
+        int ScaleRatio_old = configIniRead->value("/VideoConfiguration/ScaleRatio").toInt();
+        bool CustRes_isEnabled_old = configIniRead->value("/VideoConfiguration/CustRes_isEnabled").toBool();
+        int CustRes_height_old = configIniRead->value("/VideoConfiguration/CustRes_height").toInt();
+        int CustRes_width_old = configIniRead->value("/VideoConfiguration/CustRes_width").toInt();
+        //=================== 比对信息 ================================
+        if(CustRes_isEnabled_old==false&&CustRes_isEnabled==false)
+        {
+            if(ScaleRatio_old!=ScaleRatio)
+            {
+                isVideoConfigChanged=true;
+            }
+            else
+            {
+                isVideoConfigChanged=false;
+            }
+        }
+        else
+        {
+            if(CustRes_isEnabled_old==true&&CustRes_isEnabled==true)
+            {
+                if(CustRes_height_old!=CustRes_height||CustRes_width_old!=CustRes_width)
+                {
+                    isVideoConfigChanged=true;
+                }
+                else
+                {
+                    isVideoConfigChanged=false;
+                }
+            }
+            else
+            {
+                isVideoConfigChanged=true;
+            }
+        }
     }
     else
     {
-        file_mkDir(SplitFramesFolderPath);
+        emit Send_video_write_VideoConfiguration(VideoConfiguration_fullPath,ScaleRatio,0,CustRes_isEnabled,CustRes_height,CustRes_width);
     }
-    QFile::remove(AudioPath);
-    video_video2images(SourceFile_fullPath,SplitFramesFolderPath,AudioPath);
-    if(!file_isFileExist(video_mp4_fullpath))//检查是否成功生成mp4
+    //=======================
+    //   检测缓存是否存在
+    //=======================
+    if(file_isFileExist(video_mp4_fullpath)&&file_isDirExist(SplitFramesFolderPath)&&file_isDirExist(ScaledFramesFolderPath)&&file_isFileExist(VideoConfiguration_fullPath))
     {
-        emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [Cannot convert video format to mp4.]"));
-        status = "Failed";
-        emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
+        if(!isVideoConfigChanged)
+        {
+            isCacheExists=true;
+            emit Send_TextBrowser_NewMessage(tr("The previous video cache file is detected and processing of the previous video cache will continue. If you want to restart processing of the current video:[")+SourceFile_fullPath+tr("], delete the cache manually."));
+        }
+        else
+        {
+            isCacheExists=false;
+            //========
+            QFile::remove(VideoConfiguration_fullPath);
+            file_DelDir(SplitFramesFolderPath);
+            QFile::remove(AudioPath);
+            if(file_ext!="mp4")
+            {
+                if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
+            }
+            //=======
+            emit Send_TextBrowser_NewMessage(tr("The previous video cache file was detected, but because you changed the settings about the video resolution, the previous cache will be deleted and processing of the video:[")+SourceFile_fullPath+tr("] will restart."));
+        }
+    }
+    else
+    {
+        isCacheExists=false;
+        //========
+        QFile::remove(VideoConfiguration_fullPath);
         file_DelDir(SplitFramesFolderPath);
         QFile::remove(AudioPath);
-        emit Send_progressbar_Add();
-        mutex_ThreadNumRunning.lock();
-        ThreadNumRunning--;
-        mutex_ThreadNumRunning.unlock();//线程数量统计-1s
-        return 0;//如果启用stop位,则直接return
+        if(file_ext!="mp4")
+        {
+            if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
+        }
+        //========
+    }
+    if(!isCacheExists)
+    {
+        //============================== 拆分 ==========================================
+        if(file_isDirExist(SplitFramesFolderPath))
+        {
+            file_DelDir(SplitFramesFolderPath);
+            file_mkDir(SplitFramesFolderPath);
+        }
+        else
+        {
+            file_mkDir(SplitFramesFolderPath);
+        }
+        QFile::remove(AudioPath);
+        video_video2images(SourceFile_fullPath,SplitFramesFolderPath,AudioPath);
+        if(!file_isFileExist(video_mp4_fullpath))//检查是否成功生成mp4
+        {
+            emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [Cannot convert video format to mp4.]"));
+            status = "Failed";
+            emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
+            emit Send_progressbar_Add();
+            mutex_ThreadNumRunning.lock();
+            ThreadNumRunning--;
+            mutex_ThreadNumRunning.unlock();//线程数量统计-1s
+            return 0;//如果启用stop位,则直接return
+        }
     }
     //============================== 扫描获取文件名 ===============================
     QStringList Frame_fileName_list = file_getFileNames_in_Folder_nofilter(SplitFramesFolderPath);
-    if(Frame_fileName_list.isEmpty())//检查是否成功拆分为帧
+    if(!isCacheExists)
     {
-        emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [Unable to split video into pictures.]"));
-        status = "Failed";
-        emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
-        file_DelDir(SplitFramesFolderPath);
-        QFile::remove(AudioPath);
-        emit Send_progressbar_Add();
-        mutex_ThreadNumRunning.lock();
-        ThreadNumRunning--;
-        mutex_ThreadNumRunning.unlock();//线程数量统计-1s
-        return 0;//如果启用stop位,则直接return
+        if(Frame_fileName_list.isEmpty())//检查是否成功拆分为帧
+        {
+            emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [Unable to split video into pictures.]"));
+            status = "Failed";
+            emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
+            emit Send_progressbar_Add();
+            mutex_ThreadNumRunning.lock();
+            ThreadNumRunning--;
+            mutex_ThreadNumRunning.unlock();//线程数量统计-1s
+            return 0;//如果启用stop位,则直接return
+        }
     }
     //============================== 放大 =======================================
     //===========建立存储放大后frame的文件夹===========
-    QString ScaledFramesFolderPath = SplitFramesFolderPath+"/scaled";
-    if(file_isDirExist(ScaledFramesFolderPath))
+    if(!isCacheExists)
     {
-        file_DelDir(ScaledFramesFolderPath);
-        file_mkDir(ScaledFramesFolderPath);
-    }
-    else
-    {
-        file_mkDir(ScaledFramesFolderPath);
+        if(file_isDirExist(ScaledFramesFolderPath))
+        {
+            file_DelDir(ScaledFramesFolderPath);
+            file_mkDir(ScaledFramesFolderPath);
+        }
+        else
+        {
+            file_mkDir(ScaledFramesFolderPath);
+        }
     }
     //==========开始放大==========================
     int InterPro_total = Frame_fileName_list.size();
@@ -158,9 +249,6 @@ int MainWindow::Anime4k_Video(int rowNum)
             {
                 Delay_msec_sleep(500);
             }
-            file_DelDir(SplitFramesFolderPath);
-            QFile::remove(AudioPath);
-            if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
             status = "Interrupted";
             emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
             emit Send_progressbar_Add();
@@ -169,9 +257,7 @@ int MainWindow::Anime4k_Video(int rowNum)
             mutex_ThreadNumRunning.unlock();//线程数量统计-1s
             return 0;//如果启用stop位,则直接return
         }
-        //====
         Sub_video_ThreadNumRunning++;
-        //====
         QString Frame_fileName = Frame_fileName_list.at(i);
         Sub_Thread_info["Frame_fileName"]=Frame_fileName;
         QtConcurrent::run(this,&MainWindow::Anime4k_Video_scale,Sub_Thread_info,&Sub_video_ThreadNumRunning,&Frame_failed);
@@ -184,9 +270,6 @@ int MainWindow::Anime4k_Video(int rowNum)
             emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [Unable to scale all frames.]"));
             status = "Failed";
             emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
-            file_DelDir(SplitFramesFolderPath);
-            QFile::remove(AudioPath);
-            if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
             emit Send_progressbar_Add();
             mutex_ThreadNumRunning.lock();
             ThreadNumRunning--;
@@ -194,21 +277,17 @@ int MainWindow::Anime4k_Video(int rowNum)
             return 0;//如果启用stop位,则直接return
         }
     }
-    //=======================================================
-    while (Sub_video_ThreadNumRunning>0)
+    while (Sub_video_ThreadNumRunning!=0)
     {
         Delay_msec_sleep(500);
     }
     //================ 扫描放大后的帧文件数量,判断是否放大成功 =======================
     QStringList Frame_fileName_list_scaled = file_getFileNames_in_Folder_nofilter(ScaledFramesFolderPath);
-    if(Frame_fileName_list.count()!=Frame_fileName_list_scaled.count())
+    if(Frame_fileName_list_scaled.count()<Frame_fileName_list.count())
     {
         emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [Unable to scale all frames.]"));
         status = "Failed";
         emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
-        file_DelDir(SplitFramesFolderPath);
-        if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
-        QFile::remove(AudioPath);
         emit Send_progressbar_Add();
         mutex_ThreadNumRunning.lock();
         ThreadNumRunning--;
@@ -223,7 +302,7 @@ int MainWindow::Anime4k_Video(int rowNum)
     }
     else
     {
-        video_mp4_scaled_fullpath = file_path+"/"+file_name+"_waifu2x_"+QString::number(ScaleRatio,10)+"x"+"_"+file_ext+".mp4";
+        video_mp4_scaled_fullpath = file_path+"/"+file_name+"_waifu2x_"+QString::number(ScaleRatio,10)+"x_"+file_ext+".mp4";
     }
     QFile::remove(video_mp4_scaled_fullpath);
     video_images2video(video_mp4_fullpath,video_mp4_scaled_fullpath,ScaledFramesFolderPath,AudioPath,CustRes_isEnabled,CustRes_height,CustRes_width);
@@ -232,8 +311,6 @@ int MainWindow::Anime4k_Video(int rowNum)
         emit Send_TextBrowser_NewMessage(tr("Error occured when processing [")+SourceFile_fullPath+tr("]. Error: [Unable to assemble pictures into videos.]"));
         status = "Failed";
         emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
-        file_DelDir(SplitFramesFolderPath);
-        QFile::remove(AudioPath);
         emit Send_progressbar_Add();
         mutex_ThreadNumRunning.lock();
         ThreadNumRunning--;
@@ -242,11 +319,15 @@ int MainWindow::Anime4k_Video(int rowNum)
     }
     OutPutPath_Final = video_mp4_scaled_fullpath;
     //============================== 删除缓存文件 ====================================================
-    QFile::remove(AudioPath);
-    file_DelDir(SplitFramesFolderPath);
-    if(file_ext!="mp4")
+    if(!ui->checkBox_KeepVideoCache->checkState())
     {
-        if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
+        QFile::remove(VideoConfiguration_fullPath);
+        file_DelDir(SplitFramesFolderPath);
+        QFile::remove(AudioPath);
+        if(file_ext!="mp4")
+        {
+            if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
+        }
     }
     //============================= 删除原文件 & 更新filelist & 更新table status ============================
     if(DelOriginal)
@@ -259,7 +340,6 @@ int MainWindow::Anime4k_Video(int rowNum)
         {
             QFile::remove(SourceFile_fullPath);
         }
-        if(SourceFile_fullPath!=video_mp4_fullpath)QFile::remove(video_mp4_fullpath);
         status = "Finished, original file deleted";
         emit Send_Table_video_ChangeStatus_rowNumInt_statusQString(rowNum, status);
     }
@@ -277,11 +357,10 @@ int MainWindow::Anime4k_Video(int rowNum)
     }
     //============================ 更新进度条 =================================
     emit Send_progressbar_Add();
-    //=========================== 进程数量-- ==============================
+    //=========================== 更新filelist ==============================
     mutex_ThreadNumRunning.lock();
     ThreadNumRunning--;
     mutex_ThreadNumRunning.unlock();//线程数量统计-1s
-    //========
     return 0;
 }
 /*
